backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
google-generativeai
langchain-google-genai
python-dotenv
pydantic
passlib[bcrypt]
pyjwt
python-jose[cryptography]
